# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FreelanceFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List user proposals
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Generate an AI proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Proposal:
      type: object
      required: [id, jobTitle, content, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected]
        successProbability:
          type: ["number", "null"]
        tone:
          type: ["string", "null"]
        length:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        aiAnalysis:
          oneOf:
            - $ref: "#/components/schemas/ProposalAnalysis"
            - type: "null"
        createdAt:
          type: string
        updatedAt:
          type: string
    ProposalInput:
      type: object
      required: [jobTitle, jobDescription, mySkills]
      properties:
        jobId:
          type: integer
        jobTitle:
          type: string
        jobDescription:
          type: string
        mySkills:
          type: string
        budget:
          type: string
        tone:
          type: string
        length:
          type: string
          enum: [short, medium, long]
        content:
          type: string
          description: If provided, save this exact content instead of regenerating
        clientName:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        aiAnalysis:
          $ref: "#/components/schemas/ProposalAnalysis"
    ProposalAnalysis:
      type: object
      required: [scamRisk, scamReasons, budget, urgency, keywords, fitScore, fitReason]
      properties:
        clientName:
          type: ["string", "null"]
        scamRisk:
          type: string
          enum: [none, low, medium, high]
        scamReasons:
          type: array
          items:
            type: string
        budget:
          type: object
          required: [level]
          properties:
            level:
              type: string
              enum: [low, medium, high, unknown]
            estimate:
              type: ["string", "null"]
        urgency:
          type: string
          enum: [low, medium, high]
        keywords:
          type: array
          items:
            type: string
        fitScore:
          type: integer
          minimum: 0
          maximum: 100
        fitReason:
          type: string
...